refactor(wrapped-keys-lit-actions): LIT-3959 - Extract `signTransactionWithEncryptedSolanaKey` from IIFE
